Business functions
The following business functions have been added:
1. The administrator can grant/take away the administrator rights from the user
2. The customer can place an order by choosing the size of the container (number of calories) and the desired dishes in it
3. The customer can get information about how much each dish in the container will weigh and what calorie content it will be
4. The customer can view their detailed order history
5. Registration and authentication by phone number is available for couriers
6. The courier can activate/deactivate at any time to receive/not receive new orders
7. The courier can independently receive a new order
8. The courier can get detailed information about the current order and his history of delivered orders
9. The administrator can assign all available orders to couriers available at the moment (with priority according to the time of receipt of the order)
10. The customer can confirm receipt of the order
11. The courier can close the completed order only after confirmation of its receipt from the client
12. Depending on the time for which the order was completed, it is considered delivered on time or delivered late
13. The administrator can get the performance indicators of an individual courier (work speed, percentage of orders delivered on time, average number of orders delivered per day)

+++ main/src/main/java/edu/senla/controller/DishInformationController.java
@@ -3,84 +3,93 @@ package edu.senla.controller;
 import ch.qos.logback.classic.Logger;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import edu.senla.controller.controllerinterface.DishInformationControllerInterface;
-import edu.senla.dto.DishDTO;
 import edu.senla.dto.DishInformationDTO;
+import edu.senla.exeptions.ConflictBetweenData;
+import edu.senla.exeptions.NoContent;
+import edu.senla.exeptions.NotFound;
 import edu.senla.service.serviceinterface.DishInformationServiceInterface;
 import edu.senla.service.serviceinterface.DishServiceInterface;
+import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/dishesInformation")
 public class DishInformationController implements DishInformationControllerInterface {
 
-    @Autowired
-    private DishInformationServiceInterface dishInformationService;
+    private final DishInformationServiceInterface dishInformationService;
 
-    @Autowired
-    private DishServiceInterface dishService;
+    private final DishServiceInterface dishService;
 
-    @Autowired
-    private ObjectMapper jacksonObjectMapper;
+    private final ObjectMapper mapper;
 
     private final Logger LOG = (Logger) LoggerFactory.getLogger(DishInformationController.class);
+    
+    @SneakyThrows
+    @RequestMapping(method = RequestMethod.GET)
+    public ResponseEntity<String> getAllDishesInformation() {
+        LOG.info("Getting all dishes information");
+        List<DishInformationDTO> dishInformationDTOs = dishInformationService.getAllDishesInformation();
+        if (dishInformationDTOs == null || dishInformationDTOs.isEmpty()){
+            LOG.info("No dishes information found");
+            throw new NoContent();
+        }
+        String dishInformationJson = mapper.writeValueAsString(dishInformationDTOs);
+        return new ResponseEntity<String>(dishInformationJson, HttpStatus.OK);
+    }
 
     @Secured({"ROLE_ADMIN"})
     @SneakyThrows
     @RequestMapping(method = RequestMethod.POST)
     public ResponseEntity<Void> createDishInformation(@RequestBody String dishInformationJson) {
         LOG.info("Creating new dish information: {}", dishInformationJson);
-        DishInformationDTO dishInformationDTO = jacksonObjectMapper.readValue(dishInformationJson, DishInformationDTO.class);
-        int dishId = dishInformationDTO.getDishId();
-
-        try {
-            DishDTO dishDTO = dishService.readDish(dishId);
-        } catch (IllegalArgumentException exception) {
-            LOG.info("Dish for which the information is being created with id {} not found", dishId);
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        DishInformationDTO dishInformationDTO = mapper.readValue(dishInformationJson, DishInformationDTO.class);
+        if (!dishService.isDishExists(dishInformationDTO.getDishId())) {
+            LOG.info("Dish for which the information is being created with id {} not found", dishInformationDTO.getDishId());
+            throw new NotFound();
         }
-
-        dishInformationService.createDishInformation(dishId, dishInformationDTO);
+        if (dishService.isDishHasDishInformation(dishInformationDTO.getDishId())) {
+            LOG.info("Dish information for dish with id {} is already exists", dishInformationDTO.getDishId());
+            throw new ConflictBetweenData();
+        }
+        dishInformationService.createDishInformation(dishInformationDTO);
+        LOG.info("Dish information for dish with id {} successfully created", dishInformationDTO.getDishId());
         return new ResponseEntity<Void>(HttpStatus.CREATED);
     }
 
     @Secured({"ROLE_ADMIN"})
     @SneakyThrows
     @RequestMapping(value = "{id}", method = RequestMethod.GET)
-    public ResponseEntity<String> getDishInformation(@PathVariable("id") int id) {
+    public ResponseEntity<String> getDishInformation(@PathVariable("id") long id) {
         LOG.info("Getting dish information with id: {}", id);
-
-        DishInformationDTO dishInformationDTO;
-        try {
-            dishInformationDTO = dishInformationService.readDishInformation(id);
-        } catch (IllegalArgumentException exception) {
-            LOG.info("Dish information with id {} not found", id);
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (!dishInformationService.isDishInformationExists(id)) {
+            LOG.info("There is no dish information with id {}", id);
+            throw new NotFound();
         }
-
-        return new ResponseEntity<String>(jacksonObjectMapper.writeValueAsString(dishInformationDTO), HttpStatus.OK);
+        DishInformationDTO dishInformationDTO = dishInformationService.getDishInformation(id);
+        String dishInformationJson = mapper.writeValueAsString(dishInformationDTO);
+        return new ResponseEntity<String>(dishInformationJson, HttpStatus.OK);
     }
 
     @Secured({"ROLE_ADMIN"})
     @SneakyThrows
     @RequestMapping(value = "{id}", method = RequestMethod.PUT)
-    public ResponseEntity<Void> updateDishInformation(@PathVariable int id, @RequestBody String updatedDishInformationJson) {
-        LOG.info("Updating dish information: ");
-
-        try {
-            DishInformationDTO currentDishInformation = dishInformationService.readDishInformation(id);
-        } catch (IllegalArgumentException exception) {
-            LOG.info("Dish information with id {} not found", id);
-            return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<Void> updateDishInformation(@PathVariable long id, @RequestBody String updatedDishInformationJson) {
+        LOG.info("Updating dish information with id {} with new data {}: ", id, updatedDishInformationJson);
+        DishInformationDTO updatedDishInformationDTO = mapper.readValue(updatedDishInformationJson, DishInformationDTO.class);
+        if (!dishInformationService.isDishInformationExists(id)) {
+            LOG.info("There is no dish information with id {}", id);
+            throw new NotFound();
         }
-        DishInformationDTO updatedDishInformation = jacksonObjectMapper.readValue(updatedDishInformationJson, DishInformationDTO.class);
-
-        dishInformationService.updateDishInformation(id, updatedDishInformation);
+        dishInformationService.updateDishInformation(id, updatedDishInformationDTO);
+        LOG.info("Dish information with id {} successfully updated", id);
         return new ResponseEntity<Void>(HttpStatus.OK);
     }
 
@@ -88,15 +97,12 @@ public class DishInformationController implements DishInformationControllerInter
     @RequestMapping(value = "{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Void> deleteDishInformation(@PathVariable("id") int id) {
         LOG.info("Deleting dish information with id: {}", id);
-
-        try {
-            DishInformationDTO dishInformationDTO = dishInformationService.readDishInformation(id);
-        } catch (IllegalArgumentException exception) {
-            LOG.info("Unable to delete. Dish information with id {} not found", id);
-            return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+        if (!dishInformationService.isDishInformationExists(id)) {
+            LOG.info("There is no dish information with id {}", id);
+            throw new NotFound();
         }
-
         dishInformationService.deleteDishInformation(id);
+        LOG.info("Dish information with id {} successfully deleted", id);
         return new ResponseEntity<Void>(HttpStatus.OK);
     }
